% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_nucleotidi.R
\name{count_dna_nucleotides}
\alias{count_dna_nucleotides}
\title{Conta delle basi nucleotidiche}
\usage{
count_dna_nucleotides(dna)
}
\arguments{
\item{dna}{input da inserire nella console}
}
\value{
Il numero di basi nucleotidiche presenti all'interno di un frammento di DNA
}
\description{
Conta delle basi nucleotidiche
}
\examples{
dna <- "CTAGTGATGATGCTGT"
count_dna_nucleotides(dna)
}
